Make sure the log directory exists
When you follow the Readme to create a new Karafka application, it will
throw an error stating that the log dir does not exist. This directory
cannot be created by the CLI install command, because the logger needs
to be initialized before that.

This commit adds a check for the directory, and adds a spec for this as
well