Add chroot to zookeeper hosts when it is set.
The queue consumers pass the zookeeper host including the
chroot to the Poseidon::ConsumerGroup, which in turn uses
the ZK gem, which accepts chroots as part of the host url.

Extra care is taken to remove potential extra leading and
trailing slashes in the chroot. The zookeeper hosts array
is memoized, so the calculations will only be performed once