Configure load_paths in ServerCliOptions
This worked with dry-configurable < 0.11 due to a bug where original value
was passed to child classes, so mutating it in one place would make it
mutated in all of the subclasses.

This behavior was fixed in dry-configurable 0.11.0 that's why this change
is needed. It will continue to work with older versions of dry-configurable