Use the AppSignal report_error helper if available
The AppSignal for Ruby gem introduced the `Appsignal.report_error`
helper in Ruby gem version 3.10.0. This helper does what the client
already did, check if there's a current transaction or not and report
the error with `set_error` or `send_error`.

In version 4 of the Ruby gem the `report_error` helper also supports
reporting multiple errors per transaction, so it's recommended to use
this new helper.