Update Datadog metrics listener to avoid creating extra arrays
A pattern I've seen in our app while memory profiling is using `+` and
`+=` to join together arrays. Consider:

```ruby
tags = ["a:1"]
tags += ["b:2"]
````

This creates 3 arrays:
- line 1 assigning to tags (`["a:1"]`)
- the right hand side (`["b:2"]`)
- the two added together (`["a:1", "b:2"]`)

Only the last one ends up being used.

In many cases, we can reduce this to one array:

```
tags = ["a:1"]
tags.append("b:2")
```

In the casees of having another array, `concat` can be used in a
similar way to avoid creating an extra array.

Worth noting that sometimes this ends up in a slightly different order,
but for tags this _should_ be fine.